B::Deparse: Escape non-ASCII printable chars
Before this, we were escaping all non-ASCII chars if the string con-
tained any non-printable chars.  So ".\x{100}" would come out exactly
like that, but "\x{100}" would be deparsed as "Ā", causing problems if
the script has no ‘use utf8’ in scope.

(It was also misbehaving on EBCDIC, causing the existing \x
test to fail.)