Deparse require(foo()) correctly
require(foo()) deparses as require foo().  But require’s parsing is
eccentric, and does not allow require foo(), because ‘foo’ is treated
as a bareword there, and ‘require WORD’ does not allow () fol-
lowing it.

The same problem happens with do-file, which is fixed also by this
commit, since it follows the same code path.