op.c:newUNOP_AUX: Remove op_next check
This function is only called for OP_MULTIDEREF, which has no check
function.  It could conceivably be called for OP_CUSTOM, but that also
has no check function.  So this op_next check is redundant, as nothing
could possibly have set op_next at this point.  (In fact, I think it‘s
redundant for newUNOP, too, but I’ll have to look into that.)