op.c:newMETHOP: Remove fold_constants, etc.
This function is only called for OP_METHOD*.  It could conceivably be
called for OP_CUSTOM.

These operators are not eligible for constant folding, do not have
integer equivalents, do not have targets, and do not always return
scalars.  So none of the three function calls on the last line of
S_newMETHOP_internal apply.