op.c: Use scalar cx, not OA_RETSCALAR, for optimisations
OA_RETSCALAR causes the op to have scalar context applied.  If we use
the context on the op, rather than the flag on the op type, for opti-
misations, then future commits can extend these optimisations to other
cases, too.

This requires applying scalar context in newSVREF to keep \my $x
optimised (and functional, because for \my $x looks for srefgen,
not refgen).