Apply scalar cx to logops if OA_RETSCALAR
Logops always have a null-op parent, and that is what callers of
S_new_logop see.  If we flag it with scalar context when the logop
type has the OA_RETSCALAR flag (which means we know it will return
exactly one item), then \(state $x = 3) and \($x //= 4) will be opti-
mised to srefgen, instead of refgen.  (srefgen is a faster version of
refgen; it only handles one item.)