Fix deparsing of newline and tab with (?x)
B::Deparse was taking special care to preserve whitespace for /x pat-
terns, and avoid changing newlines to \n, etc.  But it was not tak-
ing (?x) into account.  Parsing regular expressions to apply the
right escapes to different parts of the regexp is too compli-
cated, so this commit changes it to treat all regular expres-
sions the way it was treating /x expressions.  This causes
‘cd t; ./perl TEST -deparse run/switchM.t’ to pass.