Eviscerate concise-xs.t
This test script constantly fails due to unrelated changes elsewhere,
and it takes a long time to run.  If we are just testing B::Concise’s
ability to distinguish types of subs, we don’t need to test so many.

For the sake of sanity, this commit rewrites the test to use eight
known subs of various persuasions.