t/run/locale.t: Rmv duplicate closing of STDERR
STDERR is now closed during these tests, so no need to do it again.  And
if you wanted to see STDERR for some reason, you had to remember to fix
things in two places.  This also sets PERL_BADLANG to suppress some of
the things that we were closing STDERR to avoid printing.