Fix garbage output from ‘my $é’
$ echo 'my $é' | iconv -f utf-8 -t iso-8859-1 | ./perl -Ilib -CS
Can't use global $^ÿ¿¿¿¿¿¿¿¿¾© in "my" at - line 1, at end of line

In trying to turn the é into a ^X sequence, alloc_my does toCTRL()
(^64) on it, where the character is treated as a signed inte-
ger, i.e., -23 ^ 64, or -87.  In sv_vcatpvfn_flags, it gets
cast to a UV, so it becomes 18446744073709551529 on 64-bit
platforms. Encoded in ‘extended’ UTF8, that becomes
<FF><80><8F><BF><BF><BF><BF><BF><BF><BF><BF><BE><A9> (assuming I cop-
ied and pasted it from less correctly), which only perl recognizes as
chr 18446744073709551529.

We should not be trying to do $^X on non-ASCII characters.