Revert ‘alternate way to figure out prototypes’
This reverts commit 1d38190f803f2a83884e9591276436394df55dd4.

$ ./perl -Ilib -MO=-q,Deparse -e 'BEGIN { require "t/test.pl"; is("foo","foo") }'
sub BEGIN {
    require 't/test.pl';
    is 'foo', 'foo';
}
-e syntax OK

The resulting code doesn’t even compile, because it lacks parentheses
around the arguments to ‘is’, and the function does not exist yet when
that statement is compiled.

The logic that makes ‘is’ lose its parentheses was added by
this commit:

commit 1d38190f803f2a83884e9591276436394df55dd4
Author: Rafael Garcia-Suarez <rgarciasuarez@gmail.com>
Date:   Tue Jan 1 23:45:21 2002 +0100

    alternate way to figure out prototypes
    Message-ID: <20020101224521.A691@rafael>

    p4raw-id: //depot/perl@14007

Previously, B::Deparse would put & on any prototyped sub call if it
were not deparsing the sub itself, because it is not possible to tell
exactly when the sub will be defined when the program loads.  (That is
still true.)

That meant is(@foo, @bar) would be deparsed as &is(@foo, @bar), which
is wrong, because the arrays are not evaluated in scalar context.
(test.pl and Test::More both have prototypes beginning with $$ for
‘is’.)  1d38190f8 was meant to fix that, and did so by respecting the
prototype if the sub exists and B::Deparse is not going to emit it
(e.g., because it’s in another file than the main program).

As of v5.21.6-321-gc65b7c4, & calls include ‘scalar’ where necessary,
so the fix from 1d38190f8 is no longer necessary.