Fix ptargrep match.
The 08_ptargrep.t test does a pattern match that appears to assume platform-specific semantics for the pattern that matches a filename.  It certainly doesn't match the filename on VMS:

not ok 1 - ptargrep shows matched file

But we know the actual filename, so the simplest solution is to just do an equality test on that, and that's what the attached patch does.

Submitted upstream at <https://rt.cpan.org/Public/Bug/Display.html?id=100903>.