[perl #123417] Allow lexer to parse \N{U+dotted.hex}
This ‘unoffical’ notation is used in stringifying regular expressions
that contain named sequences, so that qr/\N{foo}/ stringified can be
incorporated into another regular expression in a different scope and
still mean the same thing.

This also needs to work with eval "/$that_qr/".  I didn’t because the
lexer rejected this syntax.