Cherry-pick-votes file improvements
Move XSLT to the end of the file where it's sufficiently out of sight that
it doesn't need to be (almost) on one line (thanks to Aaron Crane for the
suggestion).

Add some <div> tags around the section headers to make it easy to show them
in the HTML page, since it helps make more sense of the long list of
commits to consider.