Deparse: Restore proto checks for coderef2text
This fixes #123435 (I hope).

Before v5.21.6-485-gf2279a6, B::Deparse, when deparsing sub calls,
would look in the stash for a sub with a prototype, and use that
prototype for deparsing.  (It would only do this for subs that it
itself is not going to deparse.)  That caused this to deparse
incorrectly:

BEGIN {
   require Test::More;
   is("foo", "foo");
}

Because of the presence of the prototyped &main::is by the time
B::Deparse is invoked, it would deparse the is() sub call without
parentheses, resulting in invalid code.

So I changed it to trust only those subs that it has already deparsed.

Now, with coderef2text, B::Deparse doesn’t deparse any subs except the
one passed to it.  So *all* prototype sub calls turn into &foo, break-
ing CPAN tests that expect otherwise.  While such tests could be con-
sidered too sensitive, I don’t think that this is necessarily a sane
default for coderef2text.

Ideally we should provide the option to tell coderef2text which
subs to trust, the default being to trust all of them.  But for
now I will just implement that default, which means reverting
v5.21.6-485-gf2279a6 *for coderef2text*, but not when deparsing the
main program.