[perl #123020] Scalar cx for lhs of void (...)x...
No sane code calls x in void context (except at the end of a subrou-
tine), but we have to handle it anyway.

Previously, the left-hand parenthesized operand to (...)x... would be
evaluated in the context in which the current sub was called.  I.e.,
if the last statement in the current sub is called in last context,
then then lhs of (...)x... a hundred lines earlier (say we have a
big sub) is called in list context, even though it is completely
unrelated.

Since the left-hand operand could be called in any of the three con-
texts when x itself was in void context, I could choose whatever I
wanted when making it consistent.  Scalar context makes the most sense
to me, because x falls back to string repeat when not in list context,
and because void context makes stack handling complex.

Unfortunately, this does not fix the context if x occurs at the end of
a subroutine.  That will be a lot trickier to fix.  (Hence, the com-
plex stack handling in pp_repeat must remain.)