Storable.xs: introduce SEEN*_NN
Introduce _NN versions of the SEEN() and SEEN0() macros, that
don't bother to check whether their first arg is null.

The initial motivation for this was to to silence a compiler warning
about a conditional always being true in SEEN(&PL_sv_undef,...), but
I've also applied it to all places in the code where it's clear that the
arg can't be null, e.g.

    sv = NEWSV();
    SEEN(sv,...);

At it happens, "places where" currently equates to every usage of
SEEN/SEEN0 in the source code.