improve xs_handshake() diag message
The important part of the error message is that the binaries are
mismatched; the details of the handshake keys are an implementation
detail.

Or to put it another way, when someone mixes up their paths, getting
something like

    Fcntl.c: Invalid handshake key got 0xcf80000 needed 0xd700000, binaries are mismatched

Is a bit scary and confusing. This is hopefully (slightly) less scary:

    Fcntl.c: loadable library and perl binaries are mismatched (got handshake key 0xcf80000, needed 0xd700000)