Use GIMME_V in preference to GIMME
GIMME_V is a simpler macro that results in smaller machine code.

GIMME does not distinguish between scalar and void context.  The two
instances of GIMME == G_SCALAR that I changed (which used to match
void context too, but no longer do) are in code paths unreachable in
void context, so we don’t need to check for it.