Ignore cx of padsv for padrange optimisation
Commit v5.21.6-352-gc46871e caused all padsv ops to be marked with
scalar context.  This prevented the padrange optimisation from happen-
ing with a combination of scalars and aggregates in list context.

Adjust the padrange algorithm to ignore the distinction between list
and scalar context.  We only do this optimisation when, even in list
context, each op will push just one thing on to the stack.  (That’s
why we check OPf_REF on padav and padhv.)  So the list/scalar distinc-
tion is irrelevant.