sv_vcatpvfn_flags() avoid array bounds err
clang -fsanitize=undefined is being a bit too clever for its own good
here.

The code looks something like

    U8 vhex[VHEX_SIZE];
    ...
    v = vhex + ...;
    if (v < vend) ...

The code itself is safe, but ASan detects if you've added a value
greater than the buffer size to vhex and whines.

I've changed it so that the conditional comes first and is done in such
a way that arbitrary values can't be added to vhex.

To reproduce:

    printf "%.1000a\n", 1;

gives

    sv.c:12327:34: runtime error: index 1000 out of bounds for type 'U8 [17]'