make re_intuit_string() return correct string
Fix #123469 - Bug in split function, with utf8 strings

Each regex has two SV pointers, check_substr and check_utf8,
which hold a constant string (if any) corresponding to the longest
constant string in the regexp.

When the regex is first compiled, only one pointer is set, depending on
whether the pattern is utf8 or not; but subsequent usage of the regex can
instantiate the other pointer too. So which of the two strings
re_intuit_string() should return should be based on the UTF8ness of the
pattern, not whether check_substr is set.