utf8.c: Fix potential fold bug
The function _to_uni_fold_flags() supposedly had the ability to do
folding based on the current locale, if the correct flag is passed.
However, it didn't actually do that, returning a non-locale fold
instead.  Fortunately, this is an undocumented capability (actually, the
whole function is undocumented), and no current calls to it used the
flag.  This commit causes it to work.