pp_sys.c:pp_gmtime: Remove redundant EXTEND_MORTAL
The purpose of EXTEND_MORTAL is to extend the mortals stack only
once before pushing multiple items.  Here we are just going to push
one item, so extending it gains us nothing, as sv_2mortal is going
to do that anyway.

Moreover, we will not even be using the mortals stack if there has
been an error.