Deparse state sub defined in inner subs
When we have ‘state sub x; sub { sub x { ... } }’, the actual &x sub-
routine  is not contained in the pad of the anonymous sub.  There is
only a pad entry pointing to the outer pad.  So we have go digging for
it by following ->OUTSIDE pointers.

This is only a partial fix.  If you wrap the code above in eval
'sub{...}' and deparse it, the state sub definition is missing,
because it is not available via ->OUTSIDE pointers.  I do not know
yet whether it is possible to get to it.  The to-do test in Deparse.t
used to croak.