Fix PerlIO vtables on VMS.
f0e5c859d36afe5 broke the build because it caused the PerlIO_funcs
declarations to be const in perlio.h and EXTPERLIO in perliol.h
and on VMS, EXTPERLIO was EXTCONST which is globalref.  The compiler
considers globalref and const to be incompatible.

As a workaround, make EXTPERLIO "extern const" on VMS only.  The
whole custom global data model on VMS probably needs a rethink,
but this gets the build working again.