Remove obscure "attrib -r ..\t\*.*" command from minitest recipe
This command is not run for any other test targets, and there seems to be
no point in it, at least not now, since files under t\ are not read-only
anyway.

(It dates from the creation of the minitest target in commit 137443ea0a.
At that time, the same command was already used in the $(PERLEXE) target,
having been added there in the initial creation of win32/Makefile in
commit 0a753a7640. Needless to say, neither commit gives any details on
why the command was deemed necessary.)