Avoid glob in t/TEST on non-Windows
See ticket #123561.  The presence of ‘glob’ was causing perl to
attempt to File::Glob at compile time, before @INC is set up.

Windows needs to do file globbing on @ARGV for t/TEST to work cor-
rectly.  Windows only uses TEST for miniperl (it uses harness for
‘make test’), and under miniperl ‘glob’ doesn’t use File::Glob.  So it
is safe to use ‘glob’ on Windows.  For the sake of other systems, we
can put it in a string eval to avoid even compiling the op.